inputコンポーネント再利用のために切り分け
