+++ frontend/src/components/Inputs/config/editorConfig.ts
